Auth features implemented

+++ controllers/ApiProductController.php
@@ -2,16 +2,20 @@
 require_once "controllers/BaseController.php";
 require_once "models/RopaModel.php";
 require_once "views/ApiView.php";
+require_once "helpers/AuthHelper.php";
+
 
 class ApiProductController extends BaseController {
 
     private $ropaModel;
     private $entityAttributes;
+    private $authHelper;
 
     function __construct() {
         parent::__construct();
         $this->ropaModel = new RopaModel();
         $this->entityAttributes = $this->formatAttributesColumn($this->ropaModel->getAttributes());
+        $this->authHelper = new AuthHelper();
     }
 
     private function formatAttributesColumn(array $attributes) {
@@ -81,29 +85,40 @@ class ApiProductController extends BaseController {
     }
 
     public function deleteProduct($params = null) {
-        $id = $params[':ID'];
-        $product = $this->ropaModel->getProduct($id);
-        if($product) {
-            $this->ropaModel->deleteProduct($id);
-            $this->view->response("El producto con id = $id fue borrado", 200);
+        if($this->authHelper->isAuthorized()) {
+            $id = $params[':ID'];
+            $product = $this->ropaModel->getProduct($id);
+            if($product) {
+                $this->ropaModel->deleteProduct($id);
+                $this->view->response("El producto con id = $id fue borrado", 200);
+            }
+            else {
+                $this->view->response("El producto no existe", 404);
+            }
         }
         else {
-            $this->view->response("El producto no existe", 404);
+            $this->view->response("No autorizado", 403);
         }
+
     }
 
     public function insertProduct() {
-        $body = $this->getBody();
-        $slug = str_replace(" ", "-",trim($body->nombre));
-        if ($body) {
-            $id = $this->ropaModel->insertProduct($body->precio, $body->nombre, $body->descripcion, $body->id_coleccion_fk, $body->id_tipo_fk, $slug);
-            if ($id)
-                $this->view->response("Se inserto la tarea", 200);
+        if($this->authHelper->isAuthorized()) {
+            $body = $this->getBody();
+            if ($body) {
+                $slug = str_replace(" ", "-",trim($body->nombre));
+                $id = $this->ropaModel->insertProduct($body->precio, $body->nombre, $body->descripcion, $body->coleccion, $body->tipo, $slug);
+                if ($id)
+                    $this->view->response("Se inserto la prenda", 200);
+                else
+                    $this->view->response("Error al insertar", 500);
+            }
             else
-                $this->view->response("Error al insertar", 500);
+                $this->view->response("Invalid JSON", 400);
         }
-        else
-            $this->view->response("Invalid JSON", 400);
+        else{
+            $this->view->response("No autorizado", 403);
+        }
     }
 
 }
